Added the DECIMAL type to the RecordFieldType enum
Aligning the implementation of DECIMAL RecordFieldType with other types

Introduce RecordFieldType of DECIMAL for BigDecimal values

Adding RecordFieldType DECIMAL to ResultSetRecordSet

Introduce RecordFieldType DECIMAL into DataTypeUtils

Added missing BIGINT type to convertSimpleIfPossible method

Added missing type BIGINT to convertField method

Added missing type BIGINT to writeFieldForType

Added missing type BIGINT to parseFieldForType

Added missing type BIGINT to parseStringForType

Potentially breaking change of behaviour for Kudu Service.
Switched decimal from treatment as string to treatment as decimal

Added big decimal support for ElasticSearch

Added big decimal support for hbase processors

Added big decimal support for prometheus reporting

Added big decimal support to solr

Big decimal support in site to site reporting

Core support for big decimal

Fix issue with new toBigDecimal method

Issue where method was not correctly handling the situation where a
BigDecimal values was received as input.

Update test to reflect use of BigDecimal

Changes to unit tests to support DECIMAL type

Fix bug where NaN Double reported from jvm metrics not handled

BigDecimal implementation for mongo processor

+++ nifi-nar-bundles/nifi-extension-utils/nifi-record-utils/nifi-avro-record-utils/src/test/java/org/apache/nifi/avro/TestAvroTypeUtil.java
@@ -434,10 +434,10 @@ public class TestAvroTypeUtil {
 
         final Object resultObject = convertedMap.get("amount");
         assertNotNull(resultObject);
-        assertTrue(resultObject instanceof Double);
+        assertTrue(resultObject instanceof BigDecimal);
 
-        final Double resultDouble = (Double) resultObject;
-        assertEquals(Double.valueOf(expectedDecimalValue.doubleValue()), resultDouble);
+        final BigDecimal resultDouble = (BigDecimal) resultObject;
+        assertEquals(expectedDecimalValue, resultDouble);
     }
 
     @Test
